requirements.txt
# AI 游戏开发公司 - Python 依赖清单
# 安装命令: python -m pip install -r requirements.txt

# Web 框架和服务器
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0

# AI/LLM 相关
google-generativeai>=0.8.0
google-genai>=1.0.0
openai>=1.12.0
litellm>=1.30.0

# 数据处理
pydantic>=2.6.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# 异步和网络
aiofiles>=23.2.0
httpx>=0.27.0

# 图像处理（美术素材生成）
Pillow>=10.2.0

# 实用工具
colorama>=0.4.6
